refactor(web): promote shared module-page UI into core
