+++ web/src/core/components/Sparkline.tsx
