+++ web/src/core/components/chips.scss
